extensions-app: Don't keep ref to main window
Unlike in previous GTK version (or Clutter), destroy() no longer
breaks reference cycles but just releases the reference held by
GTK itself.

So if we hold our own reference to the window, closing it won't be
enough to be disposed, and the application won't quit.

There is no good reason for the property anyway, just create the
window on activate if necessary.